requirements.txt
django~=5.1
djangorestframework~=3.15.2
django-filter~=24.3
psycopg2-binary

pytest~=8.3.2